refactor: partial data isolation (subjects protected via jwt)

+++ src/main/java/com/devcaiqueoliveira/nexus_api/service/SubjectService.java
@@ -8,6 +8,7 @@ import com.devcaiqueoliveira.nexus_api.entity.StudySession;
 import com.devcaiqueoliveira.nexus_api.entity.Subject;
 import com.devcaiqueoliveira.nexus_api.entity.User;
 import com.devcaiqueoliveira.nexus_api.entity.enums.StudySessionStatus;
+import com.devcaiqueoliveira.nexus_api.exception.exceptions.ForbiddenActionException;
 import com.devcaiqueoliveira.nexus_api.repository.StudySessionRepository;
 import com.devcaiqueoliveira.nexus_api.repository.SubjectRepository;
 import com.devcaiqueoliveira.nexus_api.repository.UserRepository;
@@ -29,9 +30,9 @@ public class SubjectService {
     private final StudySessionRepository studySessionRepository;
 
     @Transactional
-    public SubjectResponse createSubject(SubjectRequest subjectRequest) {
+    public SubjectResponse createSubject(SubjectRequest subjectRequest, UUID loggedUserId) {
 
-        User user = userRepository.findById(subjectRequest.userId())
+        User user = userRepository.findById(loggedUserId)
                 .orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado"));
 
         Subject createdSubject = new Subject(
@@ -52,14 +53,19 @@ public class SubjectService {
                 .toList();
     }
 
-    public SubjectResponse findById(UUID id) {
-        Subject subject = subjectRepository.findById(id)
+    public SubjectResponse findById(UUID subjectId, UUID loggedUserId) {
+        Subject subject = subjectRepository.findById(subjectId)
                 .orElseThrow(() -> new EntityNotFoundException("Matéria não encontrada"));
+
+        if (!subject.getUser().getId().equals(loggedUserId)) {
+            throw new ForbiddenActionException("Você não tem permissão para acessar esta matéria");
+        }
+
         return new SubjectResponse(subject);
     }
 
     @Transactional
-    public SubjectResponse updateSubject(UUID id, SubjectUpdateRequest subjectRequest) {
+    public SubjectResponse updateSubject(UUID id, SubjectUpdateRequest subjectRequest, UUID loggedUserId) {
 
         Subject subjectToUpdate = subjectRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Matéria não encontrada"));
@@ -68,27 +74,39 @@ public class SubjectService {
         subjectToUpdate.setDescription(subjectRequest.description());
         subjectToUpdate.setTargetHours(subjectRequest.targetHours());
 
+        if (!subjectToUpdate.getUser().getId().equals(loggedUserId)) {
+            throw new ForbiddenActionException("Você não tem permissão para acessar esta matéria");
+        }
+
         Subject savedSubject = subjectRepository.save(subjectToUpdate);
 
         return new SubjectResponse(savedSubject);
     }
 
     @Transactional
-    public void deleteSubject(UUID id) {
+    public void deleteSubject(UUID subjectId, UUID loggedUserId) {
 
-        Subject subjectToDelete = subjectRepository.findById(id)
+        Subject subjectToDelete = subjectRepository.findById(subjectId)
                 .orElseThrow(() -> new EntityNotFoundException("Matéria não encontrada"));
+
+        if (!subjectToDelete.getUser().getId().equals(loggedUserId)) {
+            throw new ForbiddenActionException("Você não tem permissão para acessar esta matéria");
+        }
 
         subjectRepository.delete(subjectToDelete);
     }
 
-    public SubjectProgressResponse getSubjectProgress(UUID subjectId) {
+    public SubjectProgressResponse getSubjectProgress(UUID subjectId, UUID loggedUserId) {
 
         Subject subject = subjectRepository.findById(subjectId)
                 .orElseThrow(() -> new EntityNotFoundException("Matéria não encontrada"));
 
         List<StudySession> completedSession = studySessionRepository
                 .findAllBySubjectIdAndStatus(subjectId, StudySessionStatus.COMPLETED);
+
+        if (!subject.getUser().getId().equals(loggedUserId)) {
+            throw new ForbiddenActionException("Você não tem permissão para acessar esta matéria");
+        }
 
         long totalMinutesStudied = completedSession.stream()
                 .mapToLong(session -> Duration.between(session.getStartedAt(), session.getEndedAt()).toMinutes())
